requirements.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
black==23.12.1
certifi==2024.2.2
click==8.1.7
coverage==7.5.0
dictdiffer==0.9.0
fastapi==0.109.0
flake8==7.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
iso8601==1.1.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pluggy==1.4.0
pycodestyle==2.11.1
pydantic==2.7.1
pydantic-settings==2.1.0
pydantic_core==2.18.2
pyflakes==3.2.0
pypika-tortoise==0.1.6
pytest==7.4.4
pytest-cov==4.1.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
ruff==0.4.2
sniffio==1.3.1
starlette==0.35.1
tomlkit==0.12.4
tortoise-orm==0.20.0
typing_extensions==4.11.0
uvicorn==0.26.0
